Fix #248 

The PR restores `createMediaQuery` behaviour to leverage `px` helper to supply default units as `px` when breakpoint value is unitless.