@BrendanFDMoore 2018-09-21
I just pulled this out of the source from v2 latest and dropped it in the v3 code. Let me know if you removed this for a specific reason and were intending to solve another way.