@BrendanFDMoore 2018-09-22
```
var Benchmark = require('benchmark');
// var Fns = require('../src/index');

const num = n => typeof n === 'number' && !isNaN(n)
const px = n => num(n) ? n + 'px' : n
const createMediaQuery = n => `@media screen and (min-width: ${px(n)})`
const createMediaQueryPrev = n => `@media screen and (min-width: ${n})`


const suite = new Benchmark.Suite;

// add tests
suite.add('createMediaQuery', function() {
  createMediaQuery(680);
  // createMediaQuery('680');
  createMediaQuery('680px');
})
.add('createMediaQueryPrev', function() {
  createMediaQueryPrev(680);
  // createMediaQueryPrev('680');
  // createMediaQueryPrev('680px');
})
// add listeners
.on('cycle', function(event) {
  console.log(String(event.target));
})
.on('complete', function() {
  console.log('Fastest is ' + this.filter('fastest').map('name'));
})
// run async
.run({ 'async': true });
```
👇 
```
➜  styled-system git:(bm-fix-248-breakpoint-units) ✗ node ./scripts/benchmark.js
createMediaQuery x 15,377,376 ops/sec ±1.26% (82 runs sampled)
createMediaQueryPrev x 54,337,096 ops/sec ±0.63% (88 runs sampled)
Fastest is createMediaQueryPrev
➜  styled-system git:(bm-fix-248-breakpoint-units) ✗ node ./scripts/benchmark.js
createMediaQuery x 27,411,604 ops/sec ±0.94% (81 runs sampled)
createMediaQueryPrev x 55,912,352 ops/sec ±0.37% (84 runs sampled)
Fastest is createMediaQueryPrev
➜  styled-system git:(bm-fix-248-breakpoint-units) ✗ node ./scripts/benchmark.js
createMediaQuery x 60,470,792 ops/sec ±1.22% (83 runs sampled)
createMediaQueryPrev x 602,038,440 ops/sec ±0.39% (86 runs sampled)
Fastest is createMediaQueryPrev
```

It's certainly slower when it has to check the type and add the suffix, but also resolves the linked issue.

Alternatively, we could simply update the theming docs to be explicit that the breakpoint values should carry a unit now due to performance optimizations.