@jxnblk 2018-09-22
Thanks! I was mostly curious to see benchmarks against the currently documented ones, but this is helpful.

I'm wondering if supporting this is even worth it, since `createMediaQuery` is called in the `style` function. I don't think having to use strings with unit values is too crazy (since it's in one place in your theme)

> Alternatively, we could simply update the theming docs to be explicit that the breakpoint values should carry a unit now due to performance optimizations.

Let me think about it, but I think this option might be the better way to go