@aczekajski 2018-10-01
Since it's for performance reasons, but at the same time this is a regression, I'd like to suggest turning this  behaviour into opt-out thing. For example, my project reuses the same `breakpoints` from `theme` for other mechanisms and relies on them being numbers, not strings (for the sake of making calculations on them).

My suggestion is to have an additional `breakpointsUnit` field into `theme`. If it's not falsy, it will be simply concatenated inside `createMediaQuery`:
```
createMediaQuery = (n, unit) => `@media screen and (min-width: ${n}${unit || ''})`
```