@BrendanFDMoore 2018-10-01
Oops @jxnblk , I missed this:
> see benchmarks against the currently documented ones
where can I see the current ones and how can I provide the same numbers?

@aczekajski That could be a reasonable compromise, which I can test for perf based on the above. The extra check alone will incur _some_ extra cost, it's just a question of how much.