@BrendanFDMoore 2018-10-01
Ok, got `npm run bench` to work eventually by renaming the built cjs file to `index.js` in `/dist`.
*Benchmarks with this change:*
```
space x 457,043 ops/sec ±0.80% (88 runs sampled)
width x 1,699,514 ops/sec ±0.47% (91 runs sampled)
fontSize x 1,847,931 ops/sec ±0.71% (85 runs sampled)
color x 554,733 ops/sec ±0.78% (88 runs sampled)
style x 1,410,772 ops/sec ±0.77% (87 runs sampled)
width array x 189,809 ops/sec ±0.66% (89 runs sampled)
space array x 169,302 ops/sec ±0.62% (89 runs sampled)
fontSize array x 236,943 ops/sec ±0.54% (91 runs sampled)
color array x 200,609 ops/sec ±0.49% (89 runs sampled)
```

*Baseline before this change:*
```
space x 443,754 ops/sec ±0.66% (88 runs sampled)
width x 1,679,140 ops/sec ±0.99% (89 runs sampled)
fontSize x 1,856,617 ops/sec ±0.85% (86 runs sampled)
color x 563,012 ops/sec ±0.87% (88 runs sampled)
style x 1,428,376 ops/sec ±0.74% (88 runs sampled)
width array x 187,758 ops/sec ±0.71% (90 runs sampled)
space array x 156,278 ops/sec ±0.90% (82 runs sampled)
fontSize array x 232,481 ops/sec ±0.54% (86 runs sampled)
color array x 197,372 ops/sec ±0.60% (87 runs sampled)
```

Which suggests the cost when integrated is negligible.

@jxnblk I leave it up to you, but using these benchmarks it seems safe.