@aczekajski 2018-10-02
What code did you perform this benchmark on?

If you didn't see, I've created a pull request https://github.com/jxnblk/styled-system/pull/303 with my proposition. I turned out not to be as simple as doing what I've wrote above but still should be pretty fast.