@BrendanFDMoore 2018-10-02
@aczekajski sorry for being ambiguous. My previous comment was overall benchmark info on my change for @jxnblk . Unlike my earlier post, this measures the impact on the overall system, rather than the atomic function (which is obviously going to be slower with an extra check)