@aczekajski 2018-10-03
Since `px(n)` doesn't cause too much of a performance loss, I won't insist to use my solution. It's up to @jxnblk