@jxnblk 2018-10-10
Sorry for the delay!

@BrendanFDMoore thanks for checking the benchmarks on this! Seems like it doesn't appear to have a huge impact, so planning on merging this in and publishing a new version.

@aczekajski planning on going with this one over #303 but thanks for working on this!